renaming test modules to testing
