+++ app/src/test/java/net/squanchy/service/firebase/FirestoreEventMapperTest.kt
@@ -5,6 +5,9 @@ import net.squanchy.A_DATE
 import net.squanchy.A_TIMEZONE
 import net.squanchy.schedule.domain.view.Event
 import net.squanchy.schedule.domain.view.ExperienceLevel
+import net.squanchy.schedule.firestore.aFirestoreEvent
+import net.squanchy.schedule.firestore.aFirestorePlace
+import net.squanchy.schedule.firestore.aFirestoreTrack
 import net.squanchy.schedule.firestore.toEvent
 import net.squanchy.schedule.firestore.toPlace
 import net.squanchy.schedule.firestore.toTrack
