+++ app/src/test/java/net/squanchy/service/firebase/FirestorePlaceMapperTest.kt
@@ -1,6 +1,7 @@
 package net.squanchy.service.firebase
 
 import com.google.common.truth.Truth.assertThat
+import net.squanchy.schedule.firestore.aFirestorePlace
 import net.squanchy.schedule.firestore.toPlace
 import net.squanchy.support.lang.getOrThrow
 import org.junit.Test
