+++ app/src/test/java/net/squanchy/service/firebase/FirestoreTrackMapperTest.kt
@@ -1,6 +1,7 @@
 package net.squanchy.service.firebase
 
 import com.google.common.truth.Truth.assertThat
+import net.squanchy.schedule.firestore.aFirestoreTrack
 import net.squanchy.schedule.firestore.toTrack
 import net.squanchy.support.checksum.Checksum
 import org.junit.Before
