+++ app/src/test/java/net/squanchy/service/firebase/model/schedule/FirestoreTracksRepositoryTest.kt
@@ -4,7 +4,7 @@ import io.reactivex.Observable
 import net.squanchy.schedule.domain.view.aTrack
 import net.squanchy.schedule.firestore.FirestoreTracksRepository
 import net.squanchy.service.firebase.FirestoreDbService
-import net.squanchy.service.firebase.aFirestoreTrack
+import net.squanchy.schedule.firestore.aFirestoreTrack
 import net.squanchy.support.checksum.Checksum
 import org.junit.Before
 import org.junit.Rule
