+++ schedule/schedule-service-firestore/build.gradle
@@ -46,6 +46,8 @@ dependencies {
 
     implementation libraries.app.joda
     implementation libraries.app.arrowCore
+    implementation libraries.app.firebaseFirestore
+
 
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
