--- schedule/schedule-service-firestore/src/test/java/net/squanchy/schedule/firestore/ExampleUnitTest.java
@@ -1,18 +1,0 @@
-package net.squanchy.schedule.firestore;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-
-    @Test
-    public void addition_isCorrect() {
-        assertEquals(4, 2 + 2);
-    }
-}
