+++ settings.gradle
@@ -3,10 +3,11 @@ include ':app', ':renderthread',
     ':support:option',
     ':support:checksum',
     ':auth:auth-service',
+    ':auth:auth-service-testing',
     ':speaker:speaker-domain',
-    ':speaker:speaker-domain-test',
+    ':speaker:speaker-domain-testing',
     ':schedule:schedule-domain',
-    ':schedule:schedule-domain-test',
+    ':schedule:schedule-domain-testing',
     ':schedule:schedule-service',
     ':schedule:schedule-service-firestore',
     ':firebase:firestore-db-service'
