+++ speaker/speaker-domain-testing/.gitignore
