+++ speaker/speaker-domain-testing/build.gradle
