+++ speaker/speaker-domain-testing/src/main/java/net/squanchy/speaker/domain/view/SpeakerFixtures.kt
